+++ backend/app/Http/Requests/Admin/Announcement/UpdateAnnouncementRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Admin\Announcement;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateLiveLessonRequest extends FormRequest
+class UpdateAnnouncementRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -25,11 +25,9 @@ class UpdateLiveLessonRequest extends FormRequest
     {
         return [
             'name' => 'string|max:255',
-            'date' => 'date',
-            'url' => 'string|max:255',
-            'class_id' => 'numeric|exists:class_rooms,id',
-            'question_category_id' => 'numeric|exists:question_categories,id',
-            'company_id' => 'numeric|exists:companies,id',
+            'content' => 'string',
+            'src' => 'string',
+            'company_id' => 'nullable|numeric|exists:companies,id',
         ];
     }
 }
