--- backend/app/Http/Requests/Manager/Post/StoreAnnouncementRequest.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Manager\Post;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class StoreAnnouncementRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, mixed>
-     */
-    public function rules()
-    {
-        return [
-            'name' => 'required|string|max:255',
-            'content' => 'required|string',
-            'src' => 'nullable|string',
-            'company_id' => 'nullable',
-        ];
-    }
-}
