--- backend/app/Http/Resources/Admin/Post/AnnouncementResource.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace App\Http\Resources\Admin\Post;
-
-use Illuminate\Http\Resources\Json\JsonResource;
-
-class AnnouncementResource extends JsonResource
-{
-    /**
-     * Transform the resource into an array.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
-     */
-    public function toArray($request)
-    {
-        return [
-            'id' => $this->id,
-            'name' => $this->name,
-            'content' => $this->content,
-            'company_id' => $this->company_id,
-        ];
-    }
-}
