--- backend/app/Models/Slider.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
-
-class Slider extends Model
-{
-    use HasFactory, SoftDeletes;
-
-    protected $fillable = [
-        'name',
-        'src',
-        'company_id',
-    ];
-}
