--- backend/database/factories/SliderFactory.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace Database\Factories;
-
-use App\Models\Company;
-use Illuminate\Database\Eloquent\Factories\Factory;
-
-/**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Slider>
- */
-class SliderFactory extends Factory
-{
-    /**
-     * Define the model's default state.
-     *
-     * @return array<string, mixed>
-     */
-    public function definition()
-    {
-        return [
-            'name' => $this->faker->name,
-            'src' => $this->faker->imageUrl,
-            'company_id' => Company::factory(),
-        ];
-    }
-}
