+++ backend/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\API\Admin\AnnouncementController;
 use App\Http\Controllers\API\Admin\ClassRoomController;
 use App\Http\Controllers\API\Admin\Company\CompanyController;
 use App\Http\Controllers\API\Admin\Company\CompanyUserController;
@@ -10,8 +11,6 @@ use App\Http\Controllers\API\Admin\LessonController;
 use App\Http\Controllers\API\Admin\Payment\PaymentCouponController;
 use App\Http\Controllers\API\Admin\Payment\PaymentMethodController;
 use App\Http\Controllers\API\Admin\Payment\PaymentSettingController;
-use App\Http\Controllers\API\Admin\Post\AnnouncementController;
-use App\Http\Controllers\API\Admin\Post\SliderController;
 use App\Http\Controllers\API\Admin\Question\QuestionCatergoryController;
 use App\Http\Controllers\API\Admin\Question\QuestionController;
 use App\Http\Controllers\API\Manager\Booking\BookingController;
@@ -39,6 +38,7 @@ Route::middleware(['auth:sanctum'])->group(function () {
         Route::apiResource('languages', LanguageController::class);
         Route::apiResource('class-rooms', ClassRoomController::class);
         Route::apiResource('lessons', LessonController::class);
+        Route::apiResource('announcements', AnnouncementController::class);
         Route::prefix('payment')->group(function () {
             Route::apiResource('coupons', PaymentCouponController::class);
             Route::apiResource('methods', PaymentMethodController::class);
@@ -53,22 +53,15 @@ Route::middleware(['auth:sanctum'])->group(function () {
             'questions' => QuestionController::class,
             'question/categories' => QuestionCatergoryController::class,
         ]);
-        Route::prefix('post')->group(function () {
-            Route::apiResource('announcements', AnnouncementController::class);
-            Route::apiResource('sliders', SliderController::class);
-        });
     });
 
     Route::prefix('manager')->group(function () {
         Route::apiResource('live-lessons', LiveLessonController::class);
+        Route::apiResource('announcements', \App\Http\Controllers\API\Manager\AnnouncementController::class);
         Route::apiResources([
             'bookings' => BookingController::class,
             'booking/settings' => BookingSettingController::class,
         ]);
-        Route::apiResources([
-            'post/announcements' => \App\Http\Controllers\API\Manager\Post\AnnouncementController::class,
-        ]);
-
         Route::apiResource('questions', \App\Http\Controllers\API\Manager\Question\QuestionController::class);
         Route::get('question/bugs', [\App\Http\Controllers\API\Manager\Question\QuestionController::class, 'getBugList']);
         Route::delete('question/bugs/{question}', [\App\Http\Controllers\API\Manager\Question\QuestionController::class, 'destroyBug']);
